mysite/requirements.txt
asgiref==3.7.2
chardet==5.2.0
cssselect2==0.7.0
Django==3.2.25
django-axes==6.5.1
django-classy-tags==3.0.1
django-cms==3.11.3
django-filer==2.2.7
django-formtools==2.4
django-js-asset==2.0.0
django-mptt==0.14.0
django-polymorphic==3.1.0
django-sekizai==3.0.1
django-session-timeout==0.1.0
django-sslserver==0.22
django-treebeard==4.5.1
djangocms-admin-style==1.4.0
djangocms-attributes-field==3.0.0
djangocms-bootstrap4==1.5.0
djangocms-file==2.4.0
djangocms-googlemap==1.3.0
djangocms-icon==2.1.0
djangocms-link==2.5.0
djangocms-picture==2.3.0
djangocms-snippet==2.2.0
djangocms-style==2.2.0
djangocms-text-ckeditor==5.1.5
djangocms-video==2.2.0
easy-thumbnails==2.8.5
gunicorn==21.0.0
html5lib==1.1
importlib-metadata==6.7.0
lxml==5.2.2
packaging==24.0
Pillow==9.5.0
pytz==2024.1
reportlab==4.2.2
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
Unidecode==1.1.2
webencodings==0.5.1
whitenoise==6.5.0
zipp==3.15.0
